+++ packages/@expo/prebuild-config/src/plugins/unversioned/expo-splash-screen/withIosSplashColors.ts
@@ -18,7 +18,10 @@ export const withIosSplashColors: ConfigPlugin<IOSSplashConfig> = (config, splas
   return withDangerousMod(config, [
     'ios',
     async (config) => {
-      const iosNamedProjectRoot = IOSConfig.Paths.getSourceRoot(config.modRequest.projectRoot);
+      const iosNamedProjectRoot = IOSConfig.Paths.getSourceRoot(
+        config.modRequest.projectRoot,
+        config.modRequest.platform
+      );
 
       await configureColorAssets({
         iosNamedProjectRoot,